+++ Source/EventFlow.MsSql/Integrations/TableParameter.cs
@@ -44,7 +44,7 @@ namespace EventFlow.MsSql.Integrations
         // The PropertyInfos and SqlMetaDatas static fields are dependent on the TRow type
         private static readonly Dictionary<SqlDbType, Action<SqlDataRecord, int, object>> SqlDataRecordSetters = new Dictionary<SqlDbType, Action<SqlDataRecord, int, object>>
             {
-                {SqlDbType.Text, (r, i, o) => r.SetString(i, (string)o)},
+                {SqlDbType.NText, (r, i, o) => r.SetString(i, (string)o)},
                 {SqlDbType.DateTimeOffset, (r, i, o) => r.SetDateTimeOffset(i, (DateTimeOffset)o)},
                 {SqlDbType.Int, (r, i, o) => r.SetInt32(i, (int)o)},
                 {SqlDbType.BigInt, (r, i, o) => r.SetInt64(i, (long)o)},
@@ -54,7 +54,7 @@ namespace EventFlow.MsSql.Integrations
             {
                 {typeof(Guid), SqlDbType.UniqueIdentifier},
                 {typeof(int), SqlDbType.Int},
-                {typeof(string), SqlDbType.Text},
+                {typeof(string), SqlDbType.NText},
                 {typeof(long), SqlDbType.BigInt},
                 {typeof(DateTime), SqlDbType.DateTime},
                 {typeof(DateTimeOffset), SqlDbType.DateTimeOffset},