+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ### New in 0.72 (not released yet)
 
-* _Nothing yet_
+* Fix: Storing events in MS SQL Server using `MsSqlEventPersistence` now correctly
+  handles non-ANSI unicode characters in strings.
 
 ### New in 0.71.3834 (released 2019-04-17)