+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 * Fix: Storing events in MS SQL Server using `MsSqlEventPersistence` now correctly
   handles non-ANSI unicode characters in strings.
+* Fix: Source link integration now works correctly
 
 ### New in 0.71.3834 (released 2019-04-17)