+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-### New in 0.72 (not released yet)
+### New in 0.73 (not released yet)
 
 * New: Configure JSON serialization: 
   ```csharp
@@ -8,6 +8,9 @@
       .AddConverter<SomeConverter>()
     )
   ```
+
+### New in 0.72.3914 (released 2019-05-28)
+
 * New: `EventFlow.TestHelpers` are now released as .NET Standard as well
 * Fix: Upgrade `EventStore.Client` to v5.0.1 and use it for both .NET Framework and .NET Core
 * Fix: Storing events in MS SQL Server using `MsSqlEventPersistence` now correctly